# Faster2: Cache the method, ivar, and arity and the ancestry memoized methods