We can limit calls to memoized_ivar_for/unmemoized_method_for and
related methods for things we've already calculated initially on the
memoize :some_method call.

Fixes prime_cache not priming inherited memoized methods, broken by #36
Fixes (un)memoize_all not clearing/priming subclass identifiers

```
flush_cache now:             121143.3 i/s
flush_cache PR 34:            16079.4 i/s - 7.53x slower
flush_cache master:            3718.0 i/s - 32.58x slower

flush_cache with args now:    43484.0 i/s
flush_cache with args PR 34:  17488.5 i/s - 2.49x slower
flush_cache with args master: 17279.3 i/s - 2.52x slower

prime_cache now:              55946.9 i/s
prime_cache PR 34:            12665.0 i/s - 4.42x slower
prime_cache master:            6057.2 i/s - 9.24x slower

prime_cache with args now:    30015.3 i/s
prime_cache with args PR 34:  14012.1 i/s - 2.14x slower
prime_cache with args master: 13914.9 i/s - 2.16x slower
```

For a class with 40 memoized methods, total allocations is also
greatly reduced where each method below is called 1,000 times:

method | PR 34 | master | now
------ | ----- | ----- | ---
prime_cache with args | 363421 | 164000 | 45004
prime_cache no args | 630535 | 168000 | 41000
flush_cache with args | 643361 | 164000 | 5000
flush_cache no args | 907607 | 168000 | 1000