@jrafanie 2015-12-10
@matthewrudy I know this makes things more complicated but at the same time, the  `memoize :name, :identifier => :student` methods and ivars were not being primed/flushed previously.  This PR fixes that along with making everything faster.  It does make it harder to test since any call to memoize will pollute a class that other tests might be using so you have to make sure you clean it up or use a unique class.