@matthewrudy 2015-12-14
But the tests pass, and you seem to be making good progress with the object count.

So let's merge it.
And I'll do a gem release in a couple of days.